Half tasks done
Create early crash for pulling the handbrake suddenly. DONE
Stop flight early by setting the throttle to 0 and pulling the handbrake DONE
Reimplement shake but based on flight speed. DONE
Possibly create the handbrake as a final save option. DONE
Ensure crashing occurs at the CORRECT place now. (Revert change in branch to use the actual distance and not temp 50%) DONE
Cancel the flight events when the TARDIS reaches 100%. DONE
Reimplement errors travelling to the end before completing the boss. DONE
Create a "Shutdown" state on consoles that aren't the main when in flight. DONE